+++ src/apis/dlog.py
@@ -19,7 +19,7 @@ for division in config.divisions:
     DIVISIONPNT[division["id"]] = int(division["point"])
 
 @app.get(f"/{config.vtc_abbr}/dlog/stats")
-async def dlogStats(request: Request, response: Response):
+async def dlogStats(request: Request, response: Response, starttime: Optional[int] = -1, endtime: Optional[int] = -1):
     rl = ratelimit(request.client.host, 'GET /dlog/stats', 60, 60)
     if rl > 0:
         response.status_code = 429
@@ -28,71 +28,140 @@ async def dlogStats(request: Request, response: Response):
     conn = newconn()
     cur = conn.cursor()
 
-    cur.execute(f"SELECT COUNT(*) FROM driver WHERE userid >= 0")
-    drivers = cur.fetchone()[0]
-    cur.execute(f"SELECT COUNT(*) FROM driver WHERE userid >= 0 AND joints >= {int(time.time())-86400}")
+    if starttime == -1 or endtime == -1:
+        starttime = 0
+        endtime = int(time.time())
+
+    ret = {}
+    # driver
+    cur.execute(f"SELECT COUNT(*) FROM driver WHERE userid >= 0 AND joints <= {endtime}")
+    totdrivers = cur.fetchone()[0]
+    totdrivers = 0 if totdrivers is None else int(totdrivers)
+    cur.execute(f"SELECT COUNT(*) FROM driver WHERE userid >= 0 AND joints >= {starttime} AND joints <= {endtime}")
     newdrivers = cur.fetchone()[0]
+    newdrivers = 0 if newdrivers is None else int(newdrivers)
 
-    cur.execute(f"SELECT COUNT(*) FROM dlog WHERE isdelivered = 1")
-    jobs = cur.fetchone()[0]
-    cur.execute(f"SELECT COUNT(*) FROM dlog WHERE isdelivered = 1 AND timestamp >= {int(time.time())-86400}")
-    newjobs = cur.fetchone()[0]
-
-    # euro profit
-    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 1")
-    europrofit = cur.fetchone()[0]
-    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 1 AND timestamp >= {int(time.time())-86400}")
-    neweuroprofit = cur.fetchone()[0]
+    ret["driver"] = {"tot": str(totdrivers), "new": str(newdrivers)}
     
-    # dollar profit
-    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 2")
-    dollarprofit = cur.fetchone()[0]
-    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 2 AND timestamp >= {int(time.time())-86400}")
+    # job / delivered / cancelled
+    item = {"job": "COUNT(*)", "distance": "SUM(distance)", "fuel": "SUM(fuel)"}
+    for key in item.keys():
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE timestamp <= {endtime}")
+        tot = cur.fetchone()[0]
+        tot = 0 if tot is None else int(tot)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE timestamp >= {starttime} AND timestamp <= {endtime}")
+        new = cur.fetchone()[0]
+        new = 0 if new is None else int(new)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE unit = 1 AND timestamp <= {endtime}")
+        totets2 = cur.fetchone()[0]
+        totets2 = 0 if totets2 is None else int(totets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newets2 = cur.fetchone()[0]
+        newets2 = 0 if newets2 is None else int(newets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE unit = 2 AND timestamp <= {endtime}")
+        totats = cur.fetchone()[0]
+        totats = 0 if totats is None else int(totats)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newats = cur.fetchone()[0]
+        newats = 0 if newats is None else int(newats)
+
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND timestamp <= {endtime}")
+        totdelivered = cur.fetchone()[0]
+        totdelivered = 0 if totdelivered is None else int(totdelivered)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newdelivered = cur.fetchone()[0]
+        newdelivered = 0 if newdelivered is None else int(newdelivered)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND unit = 1 AND timestamp <= {endtime}")
+        totdelivered_ets2 = cur.fetchone()[0]
+        totdelivered_ets2 = 0 if totdelivered_ets2 is None else int(totdelivered_ets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newdelivered_ets2 = cur.fetchone()[0]
+        newdelivered_ets2 = 0 if newdelivered_ets2 is None else int(newdelivered_ets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND unit = 2 AND timestamp <= {endtime}")
+        totdelivered_ats = cur.fetchone()[0]
+        totdelivered_ats = 0 if totdelivered_ats is None else int(totdelivered_ats)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 1 AND unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newdelivered_ats = cur.fetchone()[0]
+        newdelivered_ats = 0 if newdelivered_ats is None else int(newdelivered_ats)
+
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND timestamp <= {endtime}")
+        totcancelled = cur.fetchone()[0]
+        totcancelled = 0 if totcancelled is None else int(totcancelled)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newcancelled = cur.fetchone()[0]
+        newcancelled = 0 if newcancelled is None else int(newcancelled)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND unit = 1 AND timestamp <= {endtime}")
+        totcancelled_ets2 = cur.fetchone()[0]
+        totcancelled_ets2 = 0 if totcancelled_ets2 is None else int(totcancelled_ets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newcancelled_ets2 = cur.fetchone()[0]
+        newcancelled_ets2 = 0 if newcancelled_ets2 is None else int(newcancelled_ets2)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND unit = 2 AND timestamp <= {endtime}")
+        totcancelled_ats = cur.fetchone()[0]
+        totcancelled_ats = 0 if totcancelled_ats is None else int(totcancelled_ats)
+        cur.execute(f"SELECT {item[key]} FROM dlog WHERE isdelivered = 0 AND unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+        newcancelled_ats = cur.fetchone()[0]
+        newcancelled_ats = 0 if newcancelled_ats is None else int(newcancelled_ats)
+
+        ret[key] = {"all": {"sum": {"tot": str(tot), "new": str(new)}, \
+            "ets2": {"tot": str(totets2), "new": str(newets2)}, \
+            "ats": {"tot": str(totats), "new": str(newats)}}, \
+            "delivered": {"sum": {"tot": str(totdelivered), "new": str(newdelivered)}, \
+                    "ets2": {"tot": str(totdelivered_ets2), "new": str(newdelivered_ets2)}, \
+                    "ats": {"tot": str(totdelivered_ats), "new": str(newdelivered_ats)}}, \
+                "cancelled": {"sum": {"tot": str(totcancelled), "new": str(newcancelled)}, \
+                    "ets2": {"tot": str(totcancelled_ets2), "new": str(newcancelled_ets2)}, \
+                    "ats": {"tot": str(totcancelled_ats), "new": str(newcancelled_ats)}}}
+
+    # profit
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 1 AND timestamp <= {endtime}")
+    toteuroprofit = cur.fetchone()[0]
+    toteuroprofit = 0 if toteuroprofit is None else int(toteuroprofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+    neweuroprofit = cur.fetchone()[0]
+    neweuroprofit = 0 if neweuroprofit is None else int(neweuroprofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 2 AND timestamp <= {endtime}")
+    totdollarprofit = cur.fetchone()[0]
+    totdollarprofit = 0 if totdollarprofit is None else int(totdollarprofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
     newdollarprofit = cur.fetchone()[0]
+    newdollarprofit = 0 if newdollarprofit is None else int(newdollarprofit)
+    allprofit = {"tot": {"euro": str(toteuroprofit), "dollar": str(totdollarprofit)}, \
+        "new": {"euro": str(neweuroprofit), "dollar": str(newdollarprofit)}}
 
-    if europrofit is None:
-        europrofit = 0
-    if dollarprofit is None:
-        dollarprofit = 0
-    if neweuroprofit is None:
-        neweuroprofit = 0
-    if newdollarprofit is None:
-        newdollarprofit = 0
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 1 AND unit = 1 AND timestamp <= {endtime}")
+    totdelivered_europrofit = cur.fetchone()[0]
+    totdelivered_europrofit = 0 if totdelivered_europrofit is None else int(totdelivered_europrofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 1 AND unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+    newdelivered_europrofit = cur.fetchone()[0]
+    newdelivered_europrofit = 0 if newdelivered_europrofit is None else int(newdelivered_europrofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 1 AND unit = 2 AND timestamp <= {endtime}")
+    totdelivered_dollarprofit = cur.fetchone()[0]
+    totdelivered_dollarprofit = 0 if totdelivered_dollarprofit is None else int(totdelivered_dollarprofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 1 AND unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+    newdelivered_dollarprofit = cur.fetchone()[0]
+    newdelivered_dollarprofit = 0 if newdelivered_dollarprofit is None else int(newdelivered_dollarprofit)
+    deliveredprofit = {"tot": {"euro": str(totdelivered_europrofit), "dollar": str(totdelivered_dollarprofit)}, \
+        "new": {"euro": str(newdelivered_europrofit), "dollar": str(newdelivered_dollarprofit)}}
+    
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 0 AND unit = 1 AND timestamp <= {endtime}")
+    totcancelled_europrofit = cur.fetchone()[0]
+    totcancelled_europrofit = 0 if totcancelled_europrofit is None else int(totcancelled_europrofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 0 AND unit = 1 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+    newcancelled_europrofit = cur.fetchone()[0]
+    newcancelled_europrofit = 0 if newcancelled_europrofit is None else int(newcancelled_europrofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 0 AND unit = 2 AND timestamp <= {endtime}")
+    totcancelled_dollarprofit = cur.fetchone()[0]
+    totcancelled_dollarprofit = 0 if totcancelled_dollarprofit is None else int(totcancelled_dollarprofit)
+    cur.execute(f"SELECT SUM(profit) FROM dlog WHERE isdelivered = 0 AND unit = 2 AND timestamp >= {starttime} AND timestamp <= {endtime}")
+    newcancelled_dollarprofit = cur.fetchone()[0]
+    newcancelled_dollarprofit = 0 if newcancelled_dollarprofit is None else int(newcancelled_dollarprofit)
+    cancelledprofit = {"tot": {"euro": str(totcancelled_europrofit), "dollar": str(totcancelled_dollarprofit)}, \
+        "new": {"euro": str(newcancelled_europrofit), "dollar": str(newcancelled_dollarprofit)}}
+    
+    ret["profit"] = {"all": allprofit, "delivered": deliveredprofit, "cancelled": cancelledprofit}
 
-    profit = {"euro": str(europrofit), "dollar": str(dollarprofit)}
-    newprofit = {"euro": str(neweuroprofit), "dollar": str(newdollarprofit)}
-
-    cur.execute(f"SELECT SUM(fuel) FROM dlog")
-    fuel = cur.fetchone()[0]
-    cur.execute(f"SELECT SUM(fuel) FROM dlog WHERE timestamp >= {int(time.time())-86400}")
-    newfuel = cur.fetchone()[0]
-    if fuel is None:
-        fuel = 0
-    if newfuel is None:
-        newfuel = 0
-    fuel = int(fuel)
-    newfuel = int(newfuel)
-
-    cur.execute(f"SELECT SUM(distance) FROM dlog")
-    distance = cur.fetchone()[0]
-    cur.execute(f"SELECT SUM(distance) FROM dlog WHERE timestamp >= {int(time.time())-86400}")
-    newdistance = cur.fetchone()[0]
-    if distance is None:
-        distance = 0
-    if newdistance is None:
-        newdistance = 0
-    distance = int(distance)
-    newdistance = int(newdistance)
-
-    cur.execute(f"SELECT COUNT(*) FROM dlog WHERE unit = 1")
-    ets2jobs = cur.fetchone()[0]
-    cur.execute(f"SELECT COUNT(*) FROM dlog WHERE unit = 2")
-    atsjobs = cur.fetchone()[0]
-
-    return {"error": False, "response": {"drivers": {"all": str(drivers), "new": str(newdrivers)}, \
-        "jobs": {"all": str(jobs), "new": str(newjobs), "ets2": str(ets2jobs), "ats": str(atsjobs)}, \
-        "profit": {"all": profit, "new": newprofit}, \
-            "fuel": {"all": str(fuel), "new": str(newfuel)}, "distance": {"all": str(distance), "new": str(newdistance)}}}
+    return {"error": False, "response": ret}
 
 @app.get(f"/{config.vtc_abbr}/dlog/chart")
 async def dlogChart(request: Request, response: Response,